Rename local-cmd to wasm-test
